pyproject.toml
[tool.poetry]
name = "langg"
version = "0.0.1"
description = ""
authors = ["BodneyC <benjcarrington@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
protobuf = "^3.17.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
